% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grvty_balancing.R
\name{grvty_balancing}
\alias{grvty_balancing}
\title{Calculate balancing factors for gravity model}
\usage{
grvty_balancing(
  od_zones,
  friction,
  zone_id,
  zone_o,
  zone_d,
  friction_o_id,
  friction_d_id,
  friction_factor,
  tolerance,
  max_iter
)
}
\arguments{
\item{od_zones}{A data frame with origins and destinations}

\item{friction}{A data frame with friction factors for each O-D pair}

\item{zone_id}{Name of ID column in od_zones}

\item{zone_o}{Name of origins column in od_zones}

\item{zone_d}{Name of destinations column in od_zones}

\item{friction_o_id}{Name of column with origin ID in friction}

\item{friction_d_id}{Name of column with destination ID in friction}

\item{friction_factor}{Name of column with friction factor in friction}

\item{tolerance}{Acceptable error (number of trips)}

\item{max_iter}{Maximum number of iterations}
}
\value{
A list of two data frames, one with the flows and one with convergence data
}
\description{
Calculate balancing factors for gravity model
}
\examples{
result <- grvty_balancing(od_zones = buffalo_zones,
                          friction = buffalo_friction,
                          zone_id = "GEOID",
                          zone_o = "hbo_prod",
                          zone_d = "hbo_attr_bal",
                          friction_o_id = "fromId",
                          friction_d_id = "toId",
                          friction_factor = "F_HBO",
                          tolerance = 5,
                          max_iter = 100000)


}
