% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_set_min_hdwy.R
\name{gtfs_set_min_hdwy}
\alias{gtfs_set_min_hdwy}
\title{Update trips and stop_times with a new set of headways}
\usage{
gtfs_set_min_hdwy(feed, route, service, new_hdwy)
}
\arguments{
\item{feed}{A GTFS object.}

\item{route}{A character string with the route ID.}

\item{service}{A character string with the service ID}

\item{new_hdwy}{A number with the new minimum headway}
}
\value{
An updated GTFS object.
}
\description{
Update trips and stop_times with a new set of headways
}
\examples{
new_feed <- edt_set_min_hdwy(feed = feed,
                             route = "19",
                             service = "1",
                             new_hdwy = 5)
}
