% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_prep_network.R
\name{osm_prep_network}
\alias{osm_prep_network}
\title{Title}
\usage{
osm_prep_network(old_file, new_file)
}
\arguments{
\item{new_file}{}
}
\value{

}
\description{
Title
}
\examples{
osm_prep_network(old_file = "existing.osm", new_file = "proposed.osm")
}
