% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_prep_network.R
\name{osm_prep_network}
\alias{osm_prep_network}
\title{Title}
\usage{
osm_prep_network(old_osm)
}
\arguments{
\item{new_file}{}
}
\value{

}
\description{
Title
}
\examples{
new_osm <- osm_prep_network(old_osm = buffalo_osm)
}
